pyproject.toml
[project]
name = "controlled-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109,<1",
    "uvicorn>=0.23,<0.31",
    "chromadb>=0.4,<0.6",
    "sentence-transformers>=2.2,<3",
    "langchain>=0.1.0,<1.1",
    "langchain-text-splitters>=0.0.1",
    "langchain-openai>=0.1.0,<0.2",
    "langchain-google-genai>=0.1.0,<0.2",
    "pydantic>=2.5,<3",
    "python-dotenv>=1.0,<2",
    "pyyaml>=6.0,<7"
]
